Gun Functionality 01
Added a mouse pick to the raycast for the gun shoot point. This uses the raycast that has already been implemented into the player character and get the point on the screen that the raycast hits. Added a delay time of .05 seconds so, the player cant spam the fire button. Created a retro gun sound in Audacity. Created a muzzle flash game object in Blender. Added both to the gun fire game script. Added a random rotation for the muzzle flash so, it looks like a random flash pattern each time the gun is fired.