Enemy attack state
Added enemy attack. Attached 3d object to enemy so, when the player gets too close, the enemy attacks. Got an attack animation to play when enemy is too close. Attached player camera to the cube to look up at enemys arm when they swing. At this point, the spawner could have a chance to reset the enemy, breaking the attack state. Worked on trying to figure out broken state. When the enemy attacks, the game, the player cant control and the enemy stops moving.