Readable letters
Fixed enemy dead body following the player by changing their follow parameters to themselves when in the ‘dead state’. Set up random spawn for item pickups. Using ammo boxes, they can now spawn randomly in locations set by the developer. Using GIMP, I downscaled an image of a letter to be used in game to match the PSX aesthetic. In Blender, created plane and imported letter image as texture. Using a canvas, panel and textmeshpro and creating a letter prefab that can be edited so, the prefab can be brought into the scene where, text can be added after for ease of use. All players controls, aside from the ‘Use’ key get locked when reading so, when the player presses ‘Use’ while reading, they can stop interacting with the letter.