Door & Keys
Continued trying to fix the ammunition pickup. Moved onto opening and closing doors with the raycast. Door plays animation and always opens away from player for ease of use in-game. Added a key that holds a string to be used with a unique door. This uses a compare string, comparing the doors ID with the Keys ID. Keys with strings are added to an array which gets looped when trying to open a locked door.