New Enemies. Gun Empty and Adding Ammo
Added another enemy type with 2 variations. This new enemy is a more traditional zombie type. They have a walk, attack, death and hit animation attached. They follow the player character. I went back to try to figure out what went wrong with my ammo pickup. I found out that it was the game manager. I was using the local variable where, it should have been a global so items like the pistol, player and letter could interact with it. Game now checks whether the player has the gun equipped. If not equipped and empty, the gun will make an empty click sound and not fire. If Player picks up the ammo, amount of bullets specified will be added to the gun, allowing the player to fire that amount of time before the gun goes back to empty.