Stalker/Follow Player & Lightswitches
Completed a stalking template. Using a chair for testing, I got the position and rotation of the player character and set the chair to appear at a distance behind the player. I had to make the game check to see if the player camera was ‘seeing’ the chairs bounds. The chair was initially floating so, I added a navmesh agent to the chair so, it would be on the ground when the player sees it but I decided to add a collider and rigidbody so, it seems like the chair was floating behind the player, then falls as the player is looking for it.
Added a suitable skybox asset pack. Using the skybox, the test scene darkens. Using this, I decided to created my useable lightswitches.