Jumpscares & Gunfire Flash
Added a spotlight from the gun that turns on and off when the gun fires to light up the environment when fired. Jumpscares added. Players raycast hits everything except for its capsule collider and windows. Jumpscare box collider set up outside window. Player has to be close enough to the window and looking out for an animated character to pop up for less than a second with jumpscare sound effect playing. The jumpscare ends and the character disappears, not repeatable.