Main Menu
Created a splash screen for the main menu. In Gimp, using layer masks, I created the splash screen using the name of the game with some rusty background to fill in the text.  Using Levels and saturation, I got the text the way I want. Created main menu scene in Unity. Added ‘Start Game’ and ‘Exit’ buttons. Start game brings me to the Test scene and the Exit button closes the application.