Dialogue choices & Flashlight Battery fix
Added multiple dialog boxes to the NPC chat. I had to make sure that when the player chooses an option that it deactivates that button object and enables the correct next text. Fixed the flashlight, when the flashlight dims to denote the battery running out, the player wasn’t able to turn off the light. Now, when the flashlights light is int this lower power state. The player can switch it off and switch it back on, going back to the lower power state that it was in.