Captain NPC
Added a captain of the ship character. Had an issue with adding the canvas from my test scene and spent some time trying to sort it out. After some testing and research, I found out that a canvas can be buggy when copy and pasting between scenes in Unity. To remedy my Canvas issue, I had to create a new canvas in the scene and from there, could drag and drop my dialogue prefab. I added an animator to the captain character who, is a child of the boat so, he also rocks back and forth with the boat.