Start Mansion exterior
Added the close and open fade in/ out from my first puzzle. When the puzzle is complete a small cinematic of the player blinking and looking towards the island plays. Disabling the players controls and mouse cursor allowed for this. When the animation is finished, the player loads into the next scene. Started outside mansion scene. Acquired necessary assets. Built terrain, water and mansion exterior.