Door Transition and Player interact message
Created a new scene to try to replicate the Resident evil style door open transitions. Added animation to doors and camera. Created sound in audacity of door creaking open. Added scene load from the doors of the mansion to the scene of the doors opening to the main hallway of the mansion. Generated lighting. Set up a player “Press e to interact message”.