Secret passage & final enemy in scene
Added a hidden staircase to the attic that can only be accessed when found the letter and key to the attic. The starcase is accessed by interacting with a picture. The picture slides away using a Tween position. The players controls get disabled as the camera moves towards the door and camera fades to black. The player is then teleported to the attic.  Brought final enemy into scene and made the spawner component activate with a box collider when the player is at a comfortable distance into the attic level.