Playtest
Started a playtest session. Added some colliders to stop the player from going out of bounds. Changed the letter UI panel to make it more readable for the player. Came across error where, if they player goes from the Main menu to a scene and back to the menu, the ‘Start game’ button seems to not work any more.