UI Popups
Change colour of ‘Where am I’ text. Added UI text for pickups. Now, when the player picks up the pistol or flashlight, text will show up on screen for a moment confirming the pickup. Added this for all keys in the house. Now, when the player picks up a key, a message appears in the top left corner of the screen to denote a key pick up. Fixed ‘Quit to desktop’ button on all scenes. It now exits the application.