Ammo and battery UI
Added a UI component that activates when the pistol is picked up. Shows ammunition count that adds when the player picks up an ammunition box. The counter goes down until the player runs out of ammunition. Also added one for the flashlight. When the light has more than 10 seconds of battery, a green indicator will show, then amber, then red until it runs out of battery.  Fixed Cupboards in kitchen.